adapt Packed pubkey and new hasher (#7)
* update

* fix java ut

* update js and so

* remove warning

* Fix mobile (#6)

* update

* update

* update makefile

* update build.sh

* update build.sh

* add condition compile

* update index.html

* update index.html

* optimize

* update package.json

---------


* clean

* update mac os and java version

* fix set panic

* fix set panic

* update so

* update dylib

* new hash

* fix ut

* update bigint version

* public some module (#8)

* export

* public zkw

* update hash

* update java version

* add base64 wasm for js

* export base64 wasm

* update poseidon

* change timestamp to u32

* update so and js

---------

+++ rust-sdk/src/common.rs
@@ -1,13 +1,13 @@
 use serde::{Deserialize, Serialize};
 
 use crate::tx::public_key_type::PublicKeyType;
-use crate::I64SerdeAsString;
+use crate::U32SerdeAsString;
 use crate::U64SerdeAsString;
 
-pub const TRANSFER_ORDER_TYPE: u8 = 4;
-pub const CONDITIONAL_TRANSFER_ORDER_TYPE: u8 = 5;
+pub const TRANSFER_ORDER_TYPE: u64 = 4;
+pub const CONDITIONAL_TRANSFER_ORDER_TYPE: u64 = 5;
 
-pub type TimestampType = i64;
+pub type TimestampType = u32;
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct OrderBase {
@@ -15,7 +15,7 @@ pub struct OrderBase {
     pub nonce: u64,
     #[serde(rename = "public_key")]
     pub public_key: PublicKeyType,
-    #[serde(rename = "expiration_timestamp", with = "I64SerdeAsString")]
+    #[serde(rename = "expiration_timestamp", with = "U32SerdeAsString")]
     pub expiration_timestamp: TimestampType,
 }
